+++ lib/_stream_readable.js
@@ -27,8 +27,15 @@ Readable.ReadableState = ReadableState;
 const EE = require('events');
 const Stream = require('stream');
 const { Buffer } = require('buffer');
-const util = require('util');
-const debug = util.debuglog('stream');
+
+let debuglog;
+function debug(...args) {
+  if (!debuglog) {
+    debuglog = require('internal/util/debuglog').debuglog('stream');
+  }
+  debuglog(...args);
+}
+
 const BufferList = require('internal/streams/buffer_list');
 const destroyImpl = require('internal/streams/destroy');
 const { getHighWaterMark } = require('internal/streams/state');
