+++ lib/internal/bootstrap/pre_execution.js
@@ -8,6 +8,8 @@ let traceEventsAsyncHook;
 function prepareMainThreadExecution() {
   setupTraceCategoryState();
 
+  setupDebugEnv();
+
   // Only main thread receives signals.
   setupSignalHandlers();
 
@@ -50,6 +52,10 @@ function initializeReport() {
       return report;
     }
   });
+}
+
+function setupDebugEnv() {
+  require('internal/util/debuglog').initializeDebugEnv(process.env.NODE_DEBUG);
 }
 
 function setupSignalHandlers() {
@@ -231,6 +237,7 @@ function loadPreloadModules() {
 }
 
 module.exports = {
+  setupDebugEnv,
   prepareMainThreadExecution,
   initializeDeprecations,
   initializeESMLoader,
