+++ lib/internal/main/worker_thread.js
@@ -4,6 +4,7 @@
 // message port.
 
 const {
+  setupDebugEnv,
   initializeDeprecations,
   initializeESMLoader,
   initializeFrozenIntrinsics,
@@ -37,6 +38,9 @@ const {
 } = require('internal/process/execution');
 
 const publicWorker = require('worker_threads');
+
+setupDebugEnv();
+
 const debug = require('util').debuglog('worker');
 
 debug(`[${threadId}] is setting up worker child environment`);
