+++ lib/internal/modules/cjs/loader.js
@@ -174,7 +174,13 @@ Object.defineProperty(Module, 'wrapper', {
   }
 });
 
-const debug = util.debuglog('module');
+let debuglog;
+function debug(...args) {
+  if (!debuglog) {
+    debuglog = require('internal/util/debuglog').debuglog('module');
+  }
+  debuglog(...args);
+}
 
 Module._debug = util.deprecate(debug, 'Module._debug is deprecated.',
                                'DEP0077');
