+++ lib/internal/worker.js
@@ -3,7 +3,7 @@
 const EventEmitter = require('events');
 const assert = require('internal/assert');
 const path = require('path');
-const util = require('util');
+
 const {
   ERR_WORKER_PATH,
   ERR_WORKER_UNSERIALIZABLE_ERROR,
@@ -45,7 +45,13 @@ const kOnCouldNotSerializeErr = Symbol('kOnCouldNotSerializeErr');
 const kOnErrorMessage = Symbol('kOnErrorMessage');
 const kParentSideStdio = Symbol('kParentSideStdio');
 
-const debug = util.debuglog('worker');
+let debuglog;
+function debug(...args) {
+  if (!debuglog) {
+    debuglog = require('internal/util/debuglog').debuglog('worker');
+  }
+  debuglog(...args);
+}
 
 class Worker extends EventEmitter {
   constructor(filename, options = {}) {
