+++ lib/internal/worker/io.js
@@ -15,7 +15,14 @@ const { threadId } = internalBinding('worker');
 const { Readable, Writable } = require('stream');
 const EventEmitter = require('events');
 const util = require('util');
-const debug = util.debuglog('worker');
+
+let debuglog;
+function debug(...args) {
+  if (!debuglog) {
+    debuglog = require('internal/util/debuglog').debuglog('worker');
+  }
+  debuglog(...args);
+}
 
 const kIncrementsPortRef = Symbol('kIncrementsPortRef');
 const kName = Symbol('kName');
