+++ lib/timers.js
@@ -42,7 +42,15 @@ const {
 const internalUtil = require('internal/util');
 const util = require('util');
 const { ERR_INVALID_CALLBACK } = require('internal/errors').codes;
-const debug = util.debuglog('timer');
+
+let debuglog;
+function debug(...args) {
+  if (!debuglog) {
+    debuglog = require('internal/util/debuglog').debuglog('timer');
+  }
+  debuglog(...args);
+}
+
 const {
   destroyHooksExist,
   // The needed emit*() functions.
