+++ lib/util.js
@@ -27,6 +27,7 @@ const {
   formatWithOptions,
   inspect
 } = require('internal/util/inspect');
+const { debuglog } = require('internal/util/debuglog');
 const {
   ERR_FALSY_VALUE_REJECTION,
   ERR_INVALID_ARG_TYPE,
@@ -51,45 +52,6 @@ function uncurryThis(func) {
 const objectToString = uncurryThis(Object.prototype.toString);
 
 let internalDeepEqual;
-
-const debugs = {};
-let debugEnvRegex = /^$/;
-if (process.env.NODE_DEBUG) {
-  let debugEnv = process.env.NODE_DEBUG;
-  debugEnv = debugEnv.replace(/[|\\{}()[\]^$+?.]/g, '\\$&')
-      .replace(/\*/g, '.*')
-      .replace(/,/g, '$|^')
-      .toUpperCase();
-  debugEnvRegex = new RegExp(`^${debugEnv}$`, 'i');
-}
-
-// Emits warning when user sets
-// NODE_DEBUG=http or NODE_DEBUG=http2.
-function emitWarningIfNeeded(set) {
-  if ('HTTP' === set || 'HTTP2' === set) {
-    process.emitWarning('Setting the NODE_DEBUG environment variable ' +
-      'to \'' + set.toLowerCase() + '\' can expose sensitive ' +
-      'data (such as passwords, tokens and authentication headers) ' +
-      'in the resulting log.');
-  }
-}
-
-function debuglog(set) {
-  set = set.toUpperCase();
-  if (!debugs[set]) {
-    if (debugEnvRegex.test(set)) {
-      const pid = process.pid;
-      emitWarningIfNeeded(set);
-      debugs[set] = function debug() {
-        const msg = exports.format.apply(exports, arguments);
-        console.error('%s %d: %s', set, pid, msg);
-      };
-    } else {
-      debugs[set] = function debug() {};
-    }
-  }
-  return debugs[set];
-}
 
 function isBoolean(arg) {
   return typeof arg === 'boolean';
